pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "IDaniil228",email = "99666220+IDaniil228@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-filter (>=25.1,<26.0)"
]

[tool.poetry]
packages = [{include = "api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
